Refactor: centralizing all API knowledge
